✨ feat: add nest gen

+++ src/utils/mkFile.ts
@@ -2,12 +2,22 @@ import fs from "fs-extra";
 import * as p from "node:path";
 import chalk from "chalk";
 
-export function mkFile(path: string, fileName: string, content: string) {
-	fs.ensureDirSync(path);
-	fs.removeSync(`${path}/${fileName}`);
-	fs.writeFileSync(`${path}/${fileName}`, content);
+export let mkFileCount = 0;
+
+export function mkFile(
+	path: string,
+	fileName: string,
+	content: string,
+	dryRun = false,
+) {
+	if (!dryRun) {
+		fs.ensureDirSync(path);
+		fs.removeSync(`${path}/${fileName}`);
+		fs.writeFileSync(`${path}/${fileName}`, content);
+	}
 	const currentDir = process.cwd();
 	const fullPath = p.join(path, fileName);
-	const relativePath = p.relative(currentDir, fullPath);
+	const relativePath = `./${p.relative(currentDir, fullPath)}`;
 	console.log(chalk.green("CREATE ") + chalk.blue(relativePath));
+	mkFileCount++;
 }
